+++ sdk/storage/azfile/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.7.1-beta.1 (Unreleased)
 
 ### Features Added
+* Exported `ShareNFSSettings` and `ShareNFSSettingsEncryptionInTransit` types.
 
 ### Breaking Changes